The world survives being closed, and so do the furnaces in it
Edited columns are written to disk when they unload and when the game exits,
and come back with their block entities. Until now everything the player did
was thrown away -- not at exit, which would have been bad enough, but at column
unload: walk two hundred blocks from a house and it was gone.

Only columns the player edited are saved. Generation is deterministic, so an
untouched column is reproduced by generating it again, and writing it out as
well would make the save grow with where the player walked rather than with
what they did. `Chunk::edited()` decides, and it is set in three places rather
than the obvious one: a block edit, a furnace tick that moved, and a click in a
furnace window -- an unlit furnace with fresh fuel ticks to no effect, so
neither of the first two would fire.

Columns go to disk whole rather than as deltas. A delta would cost 120 bytes
where this costs tens of kilobytes, and it would reconstruct the column exactly
for the generator that wrote it -- but 4d (biomes) is open and will move terrain
everywhere, at which point every delta replays onto ground that is no longer
there. Vanilla stores whole chunks for this reason; here it is scheduled rather
than hypothetical.

The container is vanilla's region file. Three deliberate departures: no
compression, because zlib would be a seventh dependency and the palette words
already are the compressed form -- the compression byte is kept in vanilla's
position so adding it later is not a format break; little-endian, for the same
one-architecture reason everything else here is; and sky light is not stored,
because it is derived, costs 0.5 ms to rebuild, and would be the largest thing
in the file. Palette entries are names rather than ids, so adding a block does
not turn saved stone into deepslate.

Loading happens inside the generation job, the one point where a single thread
owns a column and nothing else can see it. Furnaces cannot go straight in --
they live in a map the simulation walks -- so a worker parks them under a mutex
and the main thread adopts them once a frame, before the tick.

The save carries its seed and refuses to open against a different one. The whole
scheme rests on unedited columns regenerating identically, and edits from another
world would land on terrain that was never theirs with nothing able to detect it
afterwards.

`--edit X Y Z <block>` exists because `--capture` cannot dig and persistence is
the one feature whose claim is about what happens between two runs. Run it twice
against the same save; the second reports the block as already set.

Also fixes the asan preset, which had stopped building on -Wsign-conversion in
mesh/Quad.hpp. Second time this has happened, and both times ctest reported a
pass because it ran a stale binary.

350 tests, asan clean, tsan clean over twelve seconds of the running app -- run
twice, once writing and once loading.

+++ src/mesh/Quad.hpp
@@ -94,8 +94,14 @@ struct Quad {
     /// The same eight bits, named for what they mean on a fluid quad.
     constexpr u8 fluidCorners() const { return ao(); }
     /// One corner's drop, 0..3, in the corner order AO and light already use.
+    ///
+    /// `ao()` is widened explicitly rather than left to integer promotion, which
+    /// would make it an `int` and turn the mask into a signed-to-unsigned
+    /// conversion. Harmless in value, and `-Wsign-conversion` is an error here --
+    /// it broke the asan preset's *build* while debug kept passing, because debug
+    /// had no reason to recompile this header.
     constexpr u8 fluidDrop(u32 corner) const {
-        return static_cast<u8>((ao() >> (2u * corner)) & 0x3u);
+        return static_cast<u8>((static_cast<u32>(ao()) >> (2u * corner)) & 0x3u);
     }
     constexpr u16 light() const { return static_cast<u16>((packed >> 41) & 0xFFFF); }
     constexpr u16 material() const { return static_cast<u16>((packed >> 57) & 0x7F); }
